Refactor CSL parser
