+++ code-gen/parse-csl/app/Main.hs
@@ -1,9 +1,20 @@
 module Main (main) where
 
-import           Csl
-import           System.Directory   (createDirectoryIfMissing)
-import           System.Environment (getArgs)
-import           System.FilePath    (takeDirectory)
+import Csl (
+  classJs,
+  classPurs,
+  enumPurs,
+  exportListPurs,
+  funJs,
+  funPurs,
+  getClasses,
+  getEnums,
+  getFuns,
+  isCommon,
+ )
+import System.Directory (createDirectoryIfMissing)
+import System.Environment (getArgs)
+import System.FilePath (takeDirectory)
 
 main :: IO ()
 main = do
@@ -28,26 +39,29 @@ main = do
     exportsPursCode = exportListPurs nonCommonFuns classes enums
   createDirectoryIfMissing True $ takeDirectory $ exportPath <> "/"
   createDirectoryIfMissing True $ takeDirectory $ exportPath <> "/Lib/"
-  writeFile (exportPath <> "Lib.purs") $ unlines
-    [ pursLibHeader ++ exportsPursCode ++ "\n  ) where"
-    , importsCode
-    , ""
-    , "-- functions"
-    , funsPursCode
-    , ""
-    , "-- classes"
-    , ""
-    , classesPursCode
-    , "-- enums"
-    , ""
-    , enumsPursCode
-    ]
-  writeFile (exportPath <> "Lib.js") $ unlines
-    [ jsLibHeader
-    , classesJsCode
-    , funsJsCode
-    ]
+  writeFile (exportPath <> "Lib.purs") $
+    unlines
+      [ pursLibHeader ++ exportsPursCode ++ "\n  ) where"
+      , importsCode
+      , ""
+      , "-- functions"
+      , funsPursCode
+      , ""
+      , "-- classes"
+      , ""
+      , classesPursCode
+      , "-- enums"
+      , ""
+      , enumsPursCode
+      ]
+  writeFile (exportPath <> "Lib.js") $
+    unlines
+      [ jsLibHeader
+      , classesJsCode
+      , funsJsCode
+      ]
   writeFile (exportPath <> "Lib/Internal.purs") pursInternalLib
   writeFile (exportPath <> "Lib/Internal.js") jsInternalLib
 
+pursLibHeader :: [Char]
 pursLibHeader = "module Cardano.Serialization.Lib\n  ( "
